requirements.txt
streamlit==1.28.1
pandas==2.0.3
numpy==1.24.3
plotly==5.14.0
sqlalchemy==2.0.20
cx-Oracle==8.3.0
python-dotenv==1.0.0
bcrypt==4.0.1
pydantic==2.3.0
requests==2.31.0
